Store files in XDG approved location, initial multi-device support
Fit files and auth files are now stored in an the location specifed
by the XDG Base Directory specification. It uses the $XDG_CONFIG_HOME
with $HOME/.config as backup. In this directory a garmin-extractor
folder is created in which a folder for each device is created. In
this folder the fit files and auth file are stored. (fixes #2)