Make setup of GcpUploader integration easier
It's not ideal to require a user to modify code in order to get things
working.  It isn't ideal for developers either, because it pollutes the
working tree with a user-specific configuration change which then has to
be carefully excluded when submitting patches upstream.

So instead we provide a sensible default location for gupload.py, and
allow overriding it via an environment variable.